% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sink fit.R
\name{sink_fit}
\alias{sink_fit}
\title{Run segmented regressions on cell sinking data}
\usage{
sink_fit(database_rfu, breakpoints = c(20, 50), attempts = 19, export_output = FALSE,
save_as = "Sinking_Fits.csv")
}
\arguments{
\item{database_rfu}{dataframe containing the columns "Plate", "Well", "Elapsed.Time.m", and "RFU". Can be generated by the user, or by running the import functions on raw data files.}

\item{breakpoints}{initial x-axis estimates where the regression should be broken up.}

\item{attempts}{number of times the segmented regression should be run until a fit is generated. Do not change unless needed.}

\item{export_output}{whether the resulting dataframe should be exported to .csv.}

\item{save_as}{desired filename, if results are to be exported.}
}
\value{
A dataframe containing the model parameters, with each row representing a single well. The output parameters include well bottom (mean of 3 lowest points after second breakpoint), slopes 1, 2, and 3, breakpoints 1 and 2, y-intercepts on each of the slopes, and 95% upper and lower confidence intervals on each slope.
}
\description{
Fits a segmented regression to cell sinking data (square root RFU values) by well. The model parameters can then be used to calculate sinking rate.
}
\note{
The parameters generated by this function are NOT final sinking rates. You may use the sink_rate() function to generate sinking rate estimates from the parameters, or you may wish to apply your own equation.
}
\examples{
sink_parameters <- sink_fit(sinking_data)

sink_param <- sink_fit(sinking_data, breakpoints = c(10, 30))

sink_fit(sinking_data, breakpoints = c(2, 10), export_output = T, save_as = "Sink_Parameters.csv")

}
