% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sink plot.R
\name{sink_plot}
\alias{sink_plot}
\title{Plot cell sinking data with segmented regression fits}
\usage{
sink_plot(database_rfu, database_fits, size = 3, shape = 6, colour = "darkgreen",
fill = "white", shared_int = FALSE, line_1 = "red", line_2 = "blue", line_3 = "black",
export_output = FALSE, save_as = "Sinking_Plots.pdf")
}
\arguments{
\item{database_rfu}{dataframe containing the columns "Plate", "Well", "Elapsed.Time.m", and "RFU". Can be generated by the user, or by running the import functions on raw data files.}

\item{database_fits}{dataframe containing the segmented regression parameters that will be used to calculate sinking rate. Generated by running sink_fit().}

\item{size}{point size.}

\item{shape}{point shape.}

\item{colour}{point outline colour.}

\item{fill}{point fill colour.}

\item{shared_int}{whether lines for each segment should be plotted with their respective y-intercepts (the default), or if the intercept of the first segment should be used for all lines.}

\item{line_1}{line colour of first segment.}

\item{line_2}{line colour of second segment.}

\item{line_3}{line colour of third segment.}

\item{export_output}{whether the resulting dataframe should be exported to .csv.}

\item{save_as}{desired filename, if results are to be exported.}
}
\value{
Plots of square root fluorescence (sqrt(RFU)) versus elapsed time for each well on a well plate, with segmented regression lines. Either exported to .pdf or appear in the plot viewer.
}
\description{
This function plots square root fluorescence versus elapsed time with fitted segmented regression lines for each well on a well plate.
}
\examples{
sink_plot(database_rfu, database_fits, export_output = TRUE)

fitted_plots <- sink_plot(database_rfu, database_fits, line_1 = "orange")

fitted_plots <- sink_plot(database_rfu, database_fits, export_output = TRUE, save_as = "Fitted_Plots.pdf")
}
