% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sink rate.R
\name{sink_rate}
\alias{sink_rate}
\title{Calculate cell sinking rates from segmented regression parameters}
\usage{
sink_rate(database_fits, sample_depth = 3.8, export_output = FALSE, save_as = "Sinking_Rates.csv")
}
\arguments{
\item{database_fits}{dataframe containing the segmented regression parameters that will be used to calculate sinking rate. Generated by running sink_fit().}

\item{sample_depth}{height of sample in the well,in mm. Depends on the sample volume and plate size; for 300 uL samples in 96-well plates, the sample depth is 3.8 mm.}

\item{export_output}{whether the resulting dataframe should be exported to .csv.}

\item{save_as}{desired filename, if results are to be exported.}
}
\value{
A dataframe containing sinking rate estimates, with each row representing a single well. A rate will be calculated for each segment of the regression (i.e., each slope will have a corresponding sinking rate). 95% upper and lower confidence intervals on each rate are also computed.
}
\description{
Uses segmented regression parameters to estimate cell sinking rates. For a given well, a sinking rate in millimeters per minute is computed for each segment of the regression model (i.e., 3 rates per well). It is then up to the user to select which of the three sinking rates best represents the data. Sinking rates are calculated as follows: sinking rate = slope/(y-intercept - well bottom) * sample depth
}
\note{
Rates cannot be estimated without obtaining the segmented regression parameters first. There is currently no screening process in place to select the most representative sinking rate across wells/replicates, so it remains up to the user to select the rates they wish to compare across samples. It is also worth noting that sinking rates of samples with different sample depths must be estimated separately.
}
\examples{
rates <- sink_rate(sinking_fits, sample_depth = 5, export_output = T, save_as = "Sink_Rates")

sink_rate(sinking_fits, export_output = T, save_as = "Sink_Rates")

rates <- sink_rate(sinking_fits)

}
