% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot rfu.R
\name{sink_rfu}
\alias{sink_rfu}
\title{Plot fluorescence versus time by well}
\usage{
sink_rfu(database_rfu, size = 3, shape = 6, colour = "darkgreen", fill = "white",
export_output = FALSE, save_as = "Sinking_RFU.pdf")
}
\arguments{
\item{database_rfu}{dataframe containing the columns "Plate", "Well", "Elapsed.Time.m", and "RFU". Can be generated by the user, or by running the import functions on raw data files.}

\item{size}{point size.}

\item{shape}{point shape.}

\item{colour}{point outline colour.}

\item{fill}{point fill colour.}

\item{export_output}{whether resulting plots should be saved to a .pdf file. Defaults to FALSE; when FALSE, plots can be viewed within RStudio.}

\item{save_as}{desired filename, if results are to be exported.}
}
\value{
Plots of fluorescence (RFU) versus elapsed time for each well on a well plate. Either exported to .pdf or appear in the plot viewer.
}
\description{
This function plots fluorescence versus time for each well on a well plate following a sinking assay. The goal is to simply visualize the data; no data transformation or model fitting happens here.
}
\examples{
sink_rfu(Sink_Database, size = 4, shape = 1, colour = "blue", export_output = TRUE, save_as = "Sinking_Plots.pdf")
}
